backend/requirements.txt
asgiref==3.6.0
certifi==2022.12.7
Django==4.0
django-elasticsearch-dsl==7.3
django-elasticsearch-dsl-drf==0.22.5
django-filter==22.1
django-nine==0.2.7
djangorestframework==3.14.0
elasticsearch==7.17.9
elasticsearch-dsl==7.4.0
gunicorn==20.1.0
packaging==23.0
python-dateutil==2.8.2
pytz==2022.7.1
six==1.16.0
sqlparse==0.4.3
urllib3==1.26.14
